+++ CHANGES.md
@@ -7,6 +7,7 @@ Next Release (4.2.1)
 
 Features
 --------
+* [#510](https://github.com/java-native-access/jna/pull/510): Added `GetCommState`, `GetCommTimeouts` `SetCommState` and `SetCommTimeouts` to `com.sun.jna.platform.win32.Kernel32`. Added `DCB` structure to `com.sun.jna.platform.win32.WinBase` - [@MBollig](https://github.com/MBollig)
 
 Bug Fixes
 ---------